+++ mex/backend/graph/connector.py
@@ -2,6 +2,8 @@ from collections.abc import Generator, Sequence
 from typing import Annotated, Any, Literal, cast
 
 from neo4j import (
+    READ_ACCESS,
+    WRITE_ACCESS,
     Driver,
     GraphDatabase,
     NotificationDisabledCategory,
@@ -18,6 +20,7 @@ from mex.backend.graph.query import Query, QueryBuilder
 from mex.backend.graph.transform import (
     expand_references_in_search_result,
     get_error_details_from_neo4j_error,
+    get_ingest_query_for_entity_type,
     transform_edges_into_expectations_by_edge_locator,
     transform_model_into_ingest_data,
     validate_ingested_data,
@@ -101,7 +104,7 @@ class GraphConnector(BaseConnector):
             ],
             telemetry_disabled=True,
             max_connection_pool_size=settings.backend_api_parallelization,
-            max_transaction_retry_time=30.0,
+            max_transaction_retry_time=settings.graph_session_timeout,
         )
 
     def _check_connectivity_and_authentication(self) -> Result:
@@ -116,40 +119,49 @@ class GraphConnector(BaseConnector):
     def _seed_constraints(self) -> list[Result]:
         """Ensure uniqueness constraints are enabled for all entity types."""
         query_builder = QueryBuilder.get()
-        results = [
-            self.commit(
-                query_builder.create_identifier_uniqueness_constraint(
-                    node_label=class_name
+        with self.driver.session(default_access_mode=WRITE_ACCESS) as session:
+            results = [
+                self.commit(
+                    query_builder.create_identifier_uniqueness_constraint(
+                        node_label=class_name
+                    ),
+                    session=session,
                 )
-            )
-            for class_name in sorted(
-                set(EXTRACTED_MODEL_CLASSES_BY_NAME) | set(MERGED_MODEL_CLASSES_BY_NAME)
-            )
-        ]
+                for class_name in sorted(
+                    set(EXTRACTED_MODEL_CLASSES_BY_NAME)
+                    | set(MERGED_MODEL_CLASSES_BY_NAME)
+                )
+            ]
         logger.info("seeded identifier uniqueness constraints")
         return results
 
     def _seed_indices(self) -> Result:
         """Ensure there is a full text search index for all searchable fields."""
         query_builder = QueryBuilder.get()
-        result = self.commit(query_builder.fetch_full_text_search_index())
-        if (index := result.one_or_none()) and (
-            set(index["node_labels"]) != set(SEARCHABLE_CLASSES)
-            or set(index["search_fields"]) != set(SEARCHABLE_FIELDS)
-        ):
-            # only drop the index if the classes or fields have changed
-            self.commit(query_builder.drop_full_text_search_index())
-            logger.info("searchable fields changed: dropped indices")
-        result = self.commit(
-            query_builder.create_full_text_search_index(
-                node_labels=SEARCHABLE_CLASSES,
-                search_fields=SEARCHABLE_FIELDS,
-            ),
-            index_config={
-                "fulltext.eventually_consistent": True,
-                "fulltext.analyzer": "german",
-            },
-        )
+        with self.driver.session(default_access_mode=WRITE_ACCESS) as session:
+            result = self.commit(
+                query_builder.fetch_full_text_search_index(), session=session
+            )
+            if (index := result.one_or_none()) and (
+                set(index["node_labels"]) != set(SEARCHABLE_CLASSES)
+                or set(index["search_fields"]) != set(SEARCHABLE_FIELDS)
+            ):
+                # only drop the index if the classes or fields have changed
+                self.commit(
+                    query_builder.drop_full_text_search_index(), session=session
+                )
+                logger.info("searchable fields changed: dropped indices")
+            result = self.commit(
+                query_builder.create_full_text_search_index(
+                    node_labels=SEARCHABLE_CLASSES,
+                    search_fields=SEARCHABLE_FIELDS,
+                ),
+                session=session,
+                index_config={
+                    "fulltext.eventually_consistent": True,
+                    "fulltext.analyzer": "german",
+                },
+            )
         logger.info("created full text search index")
         return result
 
@@ -165,13 +177,14 @@ class GraphConnector(BaseConnector):
     def commit(
         self,
         query: Query | str,
+        /,
         session: Session | None = None,
         **parameters: Any,  # noqa: ANN401
     ) -> Result:
         """Send and commit a single graph transaction with retry configuration."""
         if session:
             return Result(session.run(str(query), parameters))
-        with self.driver.session() as closing_session:
+        with self.driver.session(default_access_mode=READ_ACCESS) as closing_session:
             return Result(closing_session.run(str(query), parameters))
 
     def _fetch_extracted_or_rule_items(  # noqa: PLR0913
@@ -533,8 +546,9 @@ class GraphConnector(BaseConnector):
 
         return result
 
-    def ingest_v2_tx(self, tx: Transaction, query: str, data_in: IngestData) -> None:
+    def ingest_v2_tx(self, tx: Transaction, data_in: IngestData) -> None:
         """Ingest a single item in a database transaction."""
+        query = get_ingest_query_for_entity_type(data_in.entityType)
         try:
             tx_result = tx.run(query, data=data_in.model_dump())
             result = Result(tx_result)
@@ -542,12 +556,19 @@ class GraphConnector(BaseConnector):
             data_out = IngestData.model_validate(result.one())
             error_details = validate_ingested_data(data_in, data_out)
             if error_details:
-                msg = f"Could not merge {data_in.nodeLabels}"
-                raise IngestionError(msg, errors=error_details)
+                msg = (
+                    f"Could not merge {data_in.entityType}"
+                    f"(stableTargetId='{data_in.stableTargetId}', ...)"
+                )
+                raise IngestionError(msg, errors=error_details, retryable=False)
         except Neo4jError as error:
             tx.rollback()
+            msg = (
+                f"{type(error).__name__} caused by {data_in.entityType}"
+                f"(stableTargetId='{data_in.stableTargetId}', ...)"
+            )
             raise IngestionError(
-                f"{type(error).__name__} caused by {data_in.nodeLabels}",  # noqa: EM102
+                msg,
                 errors=get_error_details_from_neo4j_error(data_in, error),
                 retryable=error.is_retryable(),
             ) from None
@@ -562,16 +583,17 @@ class GraphConnector(BaseConnector):
         models: Sequence[AnyExtractedModel | AnyRuleSetResponse],
     ) -> Generator[None, None, None]:
         """Ingest a list of models into the graph as nodes and connect all edges."""
-        query = str(QueryBuilder.get().merge_item_v2())
-        with self.driver.session() as session:
+        settings = BackendSettings.get()
+        with self.driver.session(default_access_mode=WRITE_ACCESS) as session:
             for model in models:
                 if isinstance(model, AnyRuleSetResponse):
                     raise NotImplementedError(AnyRuleSetResponse)
                 data_in = transform_model_into_ingest_data(model)
                 with session.begin_transaction(
-                    timeout=10.0, metadata=data_in.metadata()
+                    timeout=settings.graph_tx_timeout,
+                    metadata=data_in.metadata(),
                 ) as tx:
-                    self.ingest_v2_tx(tx, query, data_in)
+                    self.ingest_v2_tx(tx, data_in)
                 yield
 
     def ingest(
@@ -588,7 +610,7 @@ class GraphConnector(BaseConnector):
         Args:
             models: Sequence of extracted models
         """
-        with self.driver.session() as session:
+        with self.driver.session(default_access_mode=WRITE_ACCESS) as session:
             for model in models:
                 if isinstance(model, AnyRuleSetResponse):
                     for rule in (model.additive, model.subtractive, model.preventive):
@@ -622,11 +644,14 @@ class GraphConnector(BaseConnector):
         """Flush the database (only in debug mode)."""
         settings = BackendSettings.get()
         if settings.debug is True:
-            self.driver.execute_query("MATCH (n) DETACH DELETE n;")
-            for row in self.driver.execute_query("SHOW ALL CONSTRAINTS;").records:
-                self.driver.execute_query(f"DROP CONSTRAINT {row['name']};")
-            for row in self.driver.execute_query("SHOW ALL INDEXES;").records:
-                self.driver.execute_query(f"DROP INDEX {row['name']};")
+            with self.driver.session(default_access_mode=WRITE_ACCESS) as session:
+                session.run("MATCH (n) DETACH DELETE n;")
+                constraints = session.run("SHOW ALL CONSTRAINTS;")
+                for row in constraints.to_eager_result().records:
+                    session.run(f"DROP CONSTRAINT {row['name']};")
+                indexes = session.run("SHOW ALL INDEXES;")
+                for row in indexes.to_eager_result().records:
+                    session.run(f"DROP INDEX {row['name']};")
         else:
             msg = "database flush was attempted outside of debug mode"
             raise MExError(msg)